# cmd/docs/swagger.yaml
basePath: /api
definitions:
...
  handler.ProcessDatingRequestRequest:
    properties:
      action:
        type: string
    required:
    - action
    type: object
  handler.UpdateDatingRequestRequest:
    properties:
      carbon_age_error:
        type: integer
      carbon_age_value:
        type: integer
      expedition_date:
        type: string
      region:
        type: string
    required:
    - carbon_age_error
    - carbon_age_value
    - expedition_date
    - region
    type: object
...
host: localhost:8080
info:
  contact: {}
  description: Бэкенд сервис для анализа материалов и радиоуглеродного датирования.
  title: Система анализа материалов
  version: "1.0"
paths:
...
  /dating/request-materials/{requestId}/{materialId}:
    delete:
      description: Remove material from dating request
      parameters:
      - description: Request ID
        in: path
        name: requestId
        required: true
        type: integer
      - description: Material ID
        in: path
        name: materialId
        required: true
        type: integer
      responses:
        "200":
          description: Material removed successfully
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad request
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal server error
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Remove material from request
      tags:
      - Dating Requests
    put:
      consumes:
      - application/json
      description: Update material details in dating request
      parameters:
      - description: Request ID
        in: path
        name: requestId
        required: true
        type: integer
      - description: Material ID
        in: path
        name: materialId
        required: true
        type: integer
      - description: Update data
        in: body
        name: updates
        required: true
        schema:
          additionalProperties: true
          type: object
      produces:
      - application/json
      responses:
        "200":
          description: Material updated successfully
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad request
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal server error
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Update request material
      tags:
      - Dating Requests
  /dating/requests:
    get:
      description: Get list of dating requests with filtering
      parameters:
      - description: Filter by status
        in: query
        name: status
        type: string
      - description: Start date filter (YYYY-MM-DD)
        in: query
        name: start_date
        type: string
      - description: End date filter (YYYY-MM-DD)
        in: query
        name: end_date
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Success response with requests
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad request
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal server error
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Get dating requests
      tags:
      - Dating Requests
  /dating/requests/{id}:
    delete:
      description: Delete dating request by ID
      parameters:
      - description: Request ID
        in: path
        name: id
        required: true
        type: integer
      responses:
        "200":
          description: Request deleted successfully
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad request
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal server error
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Delete dating request
      tags:
      - Dating Requests
    get:
      description: Get dating request details by ID
      parameters:
      - description: Request ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: Success response with request
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad request
          schema:
            additionalProperties: true
            type: object
        "404":
          description: Request not found
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Get dating request by ID
      tags:
      - Dating Requests
    put:
      consumes:
      - application/json
      description: Update dating request data
      parameters:
      - description: Request ID
        in: path
        name: id
        required: true
        type: integer
      - description: Request update data
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/handler.UpdateDatingRequestRequest'
      produces:
      - application/json
      responses:
        "200":
          description: Request updated successfully
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad request
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal server error
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Update dating request
      tags:
      - Dating Requests
  /dating/requests/{id}/form:
    put:
      description: Form dating request for processing
      parameters:
      - description: Request ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: Request formed successfully
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad request
          schema:
            additionalProperties: true
            type: object
        "404":
          description: Request not found
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal server error
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Form dating request
      tags:
      - Dating Requests
  /dating/requests/{id}/process:
    put:
      consumes:
      - application/json
      description: Process dating request (complete/reject)
      parameters:
      - description: Request ID
        in: path
        name: id
        required: true
        type: integer
      - description: Process action
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/handler.ProcessDatingRequestRequest'
      produces:
      - application/json
      responses:
        "200":
          description: Request processed successfully
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad request
          schema:
            additionalProperties: true
            type: object
        "404":
          description: Request not found
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal server error
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Process dating request
      tags:
      - Dating Requests
...
schemes:
- http
securityDefinitions:
  BearerAuth:
    description: 'Используется для запросов через Insomnia/Postman: "Bearer <JWT>"'
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"